refactor(helpers): reorganize helpers in backward compatible way
Refs #2717
