+++ src/helpers/id-from-path-method/legacy.js
@@ -1,0 +1,3 @@
+export default function idFromPathMethodLegacy(pathName, method) {
+  return `${method.toLowerCase()}-${pathName}`;
+}
