+++ src/helpers/replace-special-chars-with-underscore.js
@@ -1,0 +1,3 @@
+const replaceSpecialCharsWithUnderscore = (operationId) => operationId.replace(/\W/gi, '_');
+
+export default replaceSpecialCharsWithUnderscore;
